Changed baseurl again :(
If baseurl is missing from _config.yml, the home page icon points to fairspectra.github.io in GitHub Pages. Works fine in local mode.
If it is set to fairspectra.net it works fine in GitHub pages, but doesn't work in local mode.
If set to "/", works locally, but not in GHP.
Trying / (not in quotes). Hopefully that fixes things in both locations.